#270 refactor training funcs
